improve picture managment

+++ back-end/controllers/post.js
@@ -38,21 +38,14 @@ exports.getOnePost = (req, res, next) => {
 // MIDDLEWARE CREATEPOST pour céer les messages
 
 exports.createPost = (req, res, next) => {
-    const userId = req.body.userID;
+    const userID = req.body.userID;
     const title = req.body.title;
-    // const gifUrl = req.body.gifUrl;
-    const gifUrl = req.body.gifUrl;
-    // const gifUrl = `${req.protocol}://${req.get("host")}/images/${req.file.filename}`;    
-    const content = req.body.content;
-
-    //console.log(imageUrl);
-    // console.log(req.file.filename);
-    
-    const createPostQuery = `INSERT INTO post VALUES (NULL, ?, ?, ?, NOW(), ?)`
-    // connect.query(createPostQuery, [userId, title, gifUrl, content], function (error, result, fields) {
-        connect.query(createPostQuery, [userId, title, gifUrl, content], function (error, result, fields) {
+    let gifUrl = `${req.protocol}://${req.get("host")}/images/${req.file.filename}`;    
+    const content = req.body.content;    
+    const createPostQuery = `INSERT INTO post VALUES (NULL, ?, ?, ?, NOW(), ?)`    
+        connect.query(createPostQuery, [userID, title, gifUrl, content], function (error, result) {
         if (error) {
-            return res.status(400).json({
+            return res.status(500).json({
                 error
             });
         }
@@ -62,57 +55,24 @@ exports.createPost = (req, res, next) => {
     });
 };
 
-
-
-// exports.createPost = (req, res, next) => {
-//     const user = decodeUid(req.headers.authorization);
-//     const { title, category } = req.body;
-//     const imageUrl = `${req.protocol}://${req.get("host")}/images/${req.file.filename}`;
-
-//     // Vérification s'il y a une image dans le body
-//     if (req.body.image === "null") {
-//         return next(new HttpError("Veuillez choisir une image", 400));
-//     }
-
-//     // Requête
-//     const string = "INSERT INTO posts (Users_id, Categories_id, title, image_url) VALUES (?, ?, ?, ? )";
-//     const inserts = [user.id, category, title, imageUrl];
-//     const sql = mysql.format(string, inserts);
-
-//     const createPost = db.query(sql, (error, post) => {
-//         if (!error) {
-//             res.status(201).json({ message: "Publication sauvegardée" });
-//         } else {
-//             return next(new HttpError("Erreur de requête, la publication n'a pas été créée", 500));
-//         }
-//     });
-// };
-
-
-
-
-
-
-
-
-
 // FIN MIDDLEWARE
 
 // MIDDLEWARE DELETEPOST pour supprimer les messages
 exports.deletePost = (req, res, next) => {
     let postID = req.params.id;
-    let userID = res.locals.userID;
+    let userID = res.locals.userID;     
+    // let gifUrl = `${req.protocol}://${req.get("host")}/images/${req.file.filename}`;    
 
     let sqlDeletePost;
     let sqlSelectPost;
 
     sqlSelectPost = "SELECT gifUrl FROM Post WHERE postID = ?";
     connect.query(sqlSelectPost, [postID], function (err, result) {
-        if (result > 0) {
-            const filename = result[0].gifUrl.split("/images/")[1];
+        if (result) {
+            const filename = result[0].gifUrl.split("/images/")[1];            
             fs.unlink(`images/${filename}`, () => { // On supprime le fichier image en amont
-                sqlDeletePost = "DELETE FROM Post WHERE postID = ?";
-                mysql.query(sqlDeletePost, [postID], function (err, result) {
+                sqlDeletePost = "DELETE FROM Post WHERE userID = ? AND postID = ?";
+                connect.query(sqlDeletePost, [userID, postID], function (err, result) {
                     if (err) {
                         return res.status(500).json(err.message);
                     }
@@ -137,15 +97,71 @@ exports.deletePost = (req, res, next) => {
 
 // MIDDLEWARE MODIFYONEPOST pour editer un post
 exports.modifyOnePost = (req, res, next) => {
+    let postID = req.params.id;
+    let title = req.body.title;
+     
+    let content = req.body.content;
 
-    connect.query(`UPDATE post SET title = '${req.body.title}', content = '${req.body.content}' WHERE post.postID = ${req.params.id}`, (error, result, field) => {
-        if (error) {
-            return res.status(400).json({
-                error
-            });
-        }
-        return res.status(200).json(result);
-    });
+    // quand es ce que ont envoi la requete json et quand cest la formdata
+    if(req.file) {
+        let gifUrl = `${req.protocol}://${req.get("host")}/images/${req.file.filename}`;
+        sqlSelectPost = "SELECT gifUrl FROM Post WHERE postID = ?";
+        connect.query(sqlSelectPost, [postID], function (err, result) {
+            if (result) {
+                const filename = result[0].gifUrl.split("/images/")[1];            
+                fs.unlink(`images/${filename}`, () => { // On supprime le fichier image en amont
+                    let sqlModifyPost = "UPDATE post SET title = ?, gifUrl = ?, content = ? WHERE postID = ?";
+                    connect.query(sqlModifyPost, [title, gifUrl, content, postID], function (error, result, field) {
+                        if (error) {
+                            return res.status(400).json({
+                                error
+                            });
+                        }
+                        return res.status(200).json(result);
+                    });
+                });
+            } else {                
+                let sqlModifyPost = "UPDATE post SET title = ?, gifUrl = ?, content = ? WHERE postID = ?";
+                connect.query(sqlModifyPost, [title, gifUrl, content, postID], function (error, result, field) {
+                    if (error) {
+                        return res.status(400).json({
+                            error
+                        });
+                    }
+                    return res.status(200).json(result);
+                });
+            }
+            if (err) {
+                return res.status(500).json(err.message);
+            }
+    
+        });
+    } else {
+        // Ici ont recoit une requete JSON
+        let sqlModifyPost = "UPDATE post SET title = ?, content = ? WHERE postID = ?";
+        connect.query(sqlModifyPost, [title, content, postID], function (error, result, field) {
+            if (error) {
+                return res.status(400).json({
+                    error
+                });
+            }
+            return res.status(200).json(result);
+        });
+    }
+
+   
+    
+    
+
+//    let sqlModifyPost = "UPDATE post SET title = ?, gifUrl = ?, content = ? WHERE postID = ?";
+//     connect.query(sqlModifyPost, [title, gifUrl, content, postID], function (error, result, field) {
+//         if (error) {
+//             return res.status(400).json({
+//                 error
+//             });
+//         }
+//         return res.status(200).json(result);
+//     });
 };
 
 
@@ -164,7 +180,7 @@ exports.createComment = (req, res, next) => {
 
 // Get all comments
 exports.getAllComments = (req, res, next) => {
-    connect.query(`SELECT user.userID, user.lastName, user.firstName, comments.commentID,comments.content, comments.userID, comments.date FROM user INNER JOIN comments ON user.userID = comments.userID WHERE comments.postID = ${req.params.id} ORDER BY comments.date DESC`,
+    connect.query(`SELECT user.userID, user.firstName, user.lastName, comments.commentID,comments.content, comments.userID, comments.date FROM user INNER JOIN comments ON user.userID = comments.userID WHERE comments.postID = ${req.params.id} ORDER BY comments.date DESC`,
         (error, result, field) => {
             if (error) {
                 return res.status(400).json({
@@ -192,41 +208,3 @@ exports.deleteComment = (req, res, next) => {
 
 // FIN MIDDLEWARE
 
-// MIDDLEWARE CREATECOMMENT pour créer des commentaires
-// exports.createComment = (req, res, next) => {
-//     const postID = req.params.id;
-//     const userID = res.locals.userID;
-//     const body = req.body.body;
-
-//     let sqlCreateComment;
-//     let values;
-
-//     sqlCreateComment = "INSERT INTO post VALUES (NULL, ?, NULL, NULL, ?, ?, NOW())";
-//     values = [userID, postID, body];
-//     connect.query(sqlCreateComment, values, function (err, result) {
-//         if (err) {
-//             return res.status(500).json(err.message);
-//         }
-//         res.status(201).json({ message: "Commentaire crée !" });
-//     });
-// };
-// FIN MIDDLEWARE
-
-// MIDDLEWARE REACTPOST pour créer une réaction sur les messages
-// exports.reactPost = (req, res, next) => {
-//     const userID = res.locals.userID;
-//     const reaction = req.body.reaction;
-//     const postID = req.params.id;
-
-//     let sqlReaction;
-//     let values;
-
-//     sqlReaction = `INSERT INTO Reaction VALUES (?, ?, ?, NOW()) ON DUPLICATE KEY UPDATE reaction = ?`;
-//     values = [userID, postID, reaction, reaction];
-//     connect.query(sqlReaction, values, function (err, result) {
-//         if (err) {
-//             return res.status(500).json(err.message);
-//         }
-//         res.status(201).json({ message: "Reaction créee !" });
-//     });
-// };
